# api/openapi.yaml
openapi: 3.1.0
info:
  title: Squyrrl API
  version: 0.1.0
  description: |
    Squyrrl（箩筐）主 API 契约。本文件描述 Phase 1 已实现的 20 个端点。

    ## 认证模型

    - 客户端用邮箱 OTP 登录得到 `access_token`（1h）+ `refresh_token`（30d）。
    - 受保护端点要求 `Authorization: Bearer <access_token>` 请求头。
    - access token 过期后用 `/auth/refresh` 旋转出新的 access；refresh token 不旋转。
    - 任何 token 都仅以 SHA-256 哈希入库；明文 token 永不持久化。

    ## 同步并发控制

    - `snippets.version` 是单调自增的乐观锁。
    - 写入端点（PATCH / DELETE）必须携带 `If-Match: "<version>"` 头，否则 400 / 409。
    - 读取端点会返回 `ETag: "<version>"` 响应头。
    - 增量同步：`GET /snippets?since=<RFC3339>` 返回 `updated_at >= since` 的碎片，按时间升序，附带 `next_since` 用于下一次请求。
    - `?include_deleted=true` 让同步流包含软删除的碎片（带 `deleted_at`），客户端据此本地下架。

    ## 数据混淆

    - `snippets.text_content` 与未来的文件 blob 在客户端用 XChaCha20-Poly1305 + 应用静态密钥混淆后再上传。
    - 服务器只见混淆字节，不解密。在 JSON 中字段以 base64 出现（Go `[]byte` 默认编码）。
  contact:
    name: Squyrrl
    url: https://squyrrl.app
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: 开发环境（本机 Docker Compose）
  - url: https://api.squyrrl.app
    description: 生产环境（占位）
...
security:
  - bearerAuth: []

# =============================================================================
# 端点
# =============================================================================
paths:
...
  /tags:
    post:
      operationId: createTag
      tags: [Tags]
      summary: 创建标签
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagCreateInput' }
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tag' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: 标签名重复
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      operationId: listTags
      tags: [Tags]
      summary: 列出标签
      description: 按 `name` 升序返回。
      responses:
        '200':
          description: 标签列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Tag' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /tags/{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPath'
    patch:
      operationId: updateTag
      tags: [Tags]
      summary: 重命名标签
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 50 }
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tag' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: 同名标签已存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      operationId: deleteTag
      tags: [Tags]
      summary: 删除标签
      description: 物理删除；通过外键级联自动清理 `snippet_tags` 关联。
      responses:
        '204': { description: 已删除 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- Snippets
  /snippets:
    post:
      operationId: createSnippet
      tags: [Snippets]
      summary: 创建碎片
      description: |
        客户端可通过 `id` 字段携带本地生成的 UUID（用于离线写）。
        `tag_ids` 与 `page_id` 必须归属当前用户，否则返回 400。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SnippetCreateInput' }
      responses:
        '201':
          description: 已创建
          headers:
            ETag:
              description: 当前 version，包在双引号里，可用作后续 If-Match
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Snippet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      operationId: listSnippets
      tags: [Snippets]
      summary: 列出 / 同步碎片
      description: |
        - 不传 `since`：浏览模式，按 `updated_at` 倒序，最多 `limit`（默认 100，上限 200）。
        - 传 `since`：同步模式，按 `updated_at` 升序，返回 `updated_at >= since` 的条目，附 `next_since` 供下次请求。
        - `include_deleted=true` 让结果包含软删除条目（`deleted_at` 非空）。
      parameters:
        - in: query
          name: page_id
          schema: { type: string, format: uuid }
        - in: query
          name: tag_id
          schema: { type: string, format: uuid }
        - in: query
          name: since
          description: RFC3339 时间戳；触发同步模式
          schema: { type: string, format: date-time }
        - in: query
          name: include_deleted
          schema: { type: boolean, default: false }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
        - in: query
          name: q
          description: |
            全文搜索关键字。Phase 1.5 仅对 `title` / `description` 做 ILIKE 单词包含匹配（大小写不敏感）。
            `text_content` 因后续将启用客户端混淆，不在服务端索引；客户端启用混淆后需要全文搜索时切换到本地索引方案。
          schema: { type: string }
      responses:
        '200':
          description: 碎片列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SnippetList' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /snippets/{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPath'
    get:
      operationId: getSnippet
      tags: [Snippets]
      summary: 拉取单个碎片
      responses:
        '200':
          description: 碎片
          headers:
            ETag:
              description: 当前 version，可作为后续 If-Match
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Snippet' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateSnippet
      tags: [Snippets]
      summary: 更新碎片（乐观锁）
      description: |
        必须携带版本号，渠道有二选一：
        - HTTP 头 `If-Match: "<version>"`（推荐）
        - body 中的 `version` 字段（兜底）

        字段存在即更新；缺省字段保持不变。`tag_ids` 为空数组表示清空所有标签。
        type / source / 等不可变字段无法通过 PATCH 修改。
      parameters:
        - in: header
          name: If-Match
          schema: { type: string }
          description: 双引号包裹的版本号，例如 `"3"`
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SnippetUpdateInput' }
      responses:
        '200':
          description: 已更新
          headers:
            ETag:
              description: 自增后的 version
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Snippet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: |
            版本不匹配（乐观锁失败）。
            服务端会自动把客户端提交的内容作为「loser 副本」入用户的「冲突收件箱」系统页面（按需建立），
            响应里包含 `loser_snippet_id` 和 `loser_page_id` 便于客户端跳转展示。
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SnippetConflictResponse' }
    delete:
      operationId: deleteSnippet
      tags: [Snippets]
      summary: 软删除碎片
      description: |
        必须携带 `If-Match` 头。删除会设置 `deleted_at = now()` 并把 `version + 1`，
        以便客户端能通过同步流（include_deleted=true）感知到此次删除。
      parameters:
        - in: header
          name: If-Match
          required: true
          schema: { type: string }
          description: 双引号包裹的版本号
      responses:
        '204': { description: 已删除 }
        '400':
          description: 缺少 If-Match 头
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: 版本不匹配
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /internal/tg/snippet:
    post:
      operationId: forwardTelegramSnippet
      tags: [Telegram]
      summary: Bot 端：以已绑定用户身份创建碎片
      description: 仅 Squyrrl Telegram Bot 调用，受 `X-Internal-Token` 头保护。服务端按 tg_user_id 找对应 user_id 创建。
      security:
        - internalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tg_user_id, snippet]
              properties:
                tg_user_id: { type: integer, format: int64 }
                snippet:
                  type: object
                  properties:
                    type: { type: string, default: normal }
                    title: { type: string }
                    description: { type: string }
                    text_content: { type: string, format: byte }
                    text_format: { type: string, default: plain }
      responses:
        '201':
          description: 碎片已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Snippet' }
        '401':
          description: 内部 token 缺失
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: tg_user_id 未绑定
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ---------------------------------------------------------------- Wallet
  /me/wallet:
    get:
      operationId: getWallet
      tags: [Wallet]
      summary: 当前用户的订阅 plan + 代币余额
      description: |
        Phase 1.5 只读视图：
        - `plan` = 用户当前 active 的会员订阅 tier；无活跃订阅时返 `"free"`。
        - `credits_balance` = `SUM(delta)` over `credits_ledger`，永远以此为准。
      responses:
        '200':
          description: 钱包视图
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wallet' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /internal/admin/credits/grant:
    post:
      operationId: grantCredits
      tags: [Wallet]
      summary: 内部：给指定用户调整 credits 余额
      description: |
        受 X-Internal-Token 头保护。供管理后台 / 支付回调 / 运营批处理调用。
        `delta` 可正可负；服务端用 advisory lock 串行化同一用户的并发写入，
        保证 `credits_ledger.balance_after` 列单调连续。
      security:
        - internalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, delta, reason]
              properties:
                user_id: { type: string, format: uuid }
                delta:
                  type: integer
                  format: int64
                  description: 正数入账，负数罚扣 / 退款冲销
                reason:
                  type: string
                  description: 流水原因，例如 'purchase' / 'admin_grant' / 'parse_uri' / 'refund'
      responses:
        '200':
          description: 流水记录已生成
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditsGrantResponse' }
        '401':
          description: 内部 token 缺失
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ---------------------------------------------------------------- Passkey (WebAuthn)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: 仅供 Bot 等内部服务调用；与主 API 共享同一个秘密
  parameters:
    UUIDPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: 请求参数无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: 未认证 / token 无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: 资源不存在或不归属当前用户
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: 人可读的错误描述（中文） }

    # ---------- 认证 ----------
...
    Tag:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TagCreateInput:
      type: object
      required: [name]
      properties:
        id:
          type: string
          format: uuid
          description: 客户端可选指定 UUID
        name: { type: string, minLength: 1, maxLength: 50 }

    # ---------- 碎片 ----------
    SnippetType:
      type: string
      enum: [uri, url, normal, special]
    SnippetTextFormat:
      type: string
      enum: [plain, markdown, code]
    SnippetSourceType:
      type: string
      enum: [device, upstream]
    Snippet:
      type: object
      required:
        - id
        - type
        - text_format
        - payload
        - source_type
        - source_data
        - version
        - tag_ids
        - created_at
        - updated_at
      properties:
        id: { type: string, format: uuid }
        page_id:
          type: [string, 'null']
          format: uuid
        type: { $ref: '#/components/schemas/SnippetType' }
        subtype:
          type: [string, 'null']
          description: type=special 时的具体子类，例如 tweet / apk / gist / youtube
        title:
          type: [string, 'null']
        description:
          type: [string, 'null']
        text_content:
          type: string
          format: byte
          description: 客户端混淆后的字节，base64 编码。服务器不解密
        text_format: { $ref: '#/components/schemas/SnippetTextFormat' }
        text_lang:
          type: [string, 'null']
          description: 仅 text_format=code 时有意义
        payload:
          type: object
          additionalProperties: true
          description: "类型相关字段，例如 url 碎片的 `{url: ...}` 或推文解析数据"
        source_type: { $ref: '#/components/schemas/SnippetSourceType' }
        source_data:
          type: object
          additionalProperties: true
          description: |
            device 时形如 `{device_id, name, platform}`；
            upstream 时形如 `{provider, author_handle, author_avatar_url, ...}`
        version:
          type: integer
          format: int64
          description: 单调自增；写入时作乐观锁
        tag_ids:
          type: array
          items: { type: string, format: uuid }
        elements:
          type: array
          items: { $ref: '#/components/schemas/Element' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at:
          type: [string, 'null']
          format: date-time
    Element:
      type: object
      required: [id, file_id, order_idx, created_at]
      properties:
        id: { type: string, format: uuid }
        file_id: { type: string, format: uuid }
        order_idx: { type: integer }
        alt_text:
          type: [string, 'null']
        created_at: { type: string, format: date-time }
    ElementInput:
      type: object
      required: [file_id]
      properties:
        file_id: { type: string, format: uuid }
        order_idx: { type: integer, default: 0 }
        alt_text: { type: string }
    SnippetCreateInput:
      type: object
      required: [type, source_type]
      properties:
        id:
          type: string
          format: uuid
          description: 客户端可选指定 UUID
        page_id:
          type: string
          format: uuid
        type: { $ref: '#/components/schemas/SnippetType' }
        subtype: { type: string }
        title: { type: string }
        description: { type: string }
        text_content: { type: string, format: byte }
        text_format: { $ref: '#/components/schemas/SnippetTextFormat' }
        text_lang: { type: string }
        payload:
          type: object
          additionalProperties: true
        source_type: { $ref: '#/components/schemas/SnippetSourceType' }
        source_data:
          type: object
          additionalProperties: true
        tag_ids:
          type: array
          items: { type: string, format: uuid }
        elements:
          type: array
          items: { $ref: '#/components/schemas/ElementInput' }
    SnippetUpdateInput:
      type: object
      description: |
        所有字段可选；缺省即不变。`version` 在 body 与 If-Match 头中至少提供一个，
        两者均存在时以 If-Match 头为准。
      properties:
        version:
          type: integer
          format: int64
          description: 兜底版本号（推荐改用 If-Match 头）
        page_id:
          type: string
          format: uuid
        title: { type: string }
        description: { type: string }
        text_content: { type: string, format: byte }
        text_format: { $ref: '#/components/schemas/SnippetTextFormat' }
        text_lang: { type: string }
        payload:
          type: object
          additionalProperties: true
        tag_ids:
          type: [array, 'null']
          description: nil = 不变；空数组 = 清空所有标签
          items: { type: string, format: uuid }
        elements:
          type: [array, 'null']
          description: nil = 不变；空数组 = 清空所有元素；非空 = 完整替换
          items: { $ref: '#/components/schemas/ElementInput' }
    SnippetList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Snippet' }
        next_since:
          type: [string, 'null']
          format: date-time
          description: |
            仅同步模式（请求带 `since`）会出现，作为下次请求的 `since` 值。
            浏览模式下省略。

    # ---------- 文件 ----------
...
    SnippetConflictResponse:
      type: object
      required: [error, current_version, loser_snippet_id, loser_page_id]
      properties:
        error: { type: string, examples: ['version conflict'] }
        current_version:
          type: integer
          format: int64
          description: 服务端当前 snippet 的实际版本，客户端应用此 version 重试
        loser_snippet_id:
          type: string
          format: uuid
          description: 客户端提交内容已作为副本入冲突收件箱的 snippet ID
        loser_page_id:
          type: string
          format: uuid
          description: 用户的「冲突收件箱」系统页面 ID

    # ---------- 归档 ----------
...
    Wallet:
      type: object
      required: [plan, credits_balance]
      properties:
        plan:
          type: string
          enum: [free, basic, standard, premium, maximum]
          description: 当前活跃订阅 tier；无活跃订阅时返 'free'
        credits_balance:
          type: integer
          format: int64
          description: SUM(delta) over credits_ledger；可为负（罚扣 / 退款冲销）
    CreditsGrantResponse:
      type: object
      required: [user_id, delta, balance_after, ledger_id]
      properties:
        user_id: { type: string, format: uuid }
        delta: { type: integer, format: int64 }
        balance_after:
          type: integer
          format: int64
          description: 该笔写入后的余额（写入瞬间的诊断字段；真实余额以 Wallet.credits_balance 为准）
        ledger_id: { type: string, format: uuid }

    # ---------- QR 扫码绑定 ----------
...